fix(api-server): resolve 7 clippy lints in migration/data_residency/regional_compliance
- Remove redundant `as i32` casts (unnecessary_cast) in migration progress %
- Remove explicit double-deref `&mut **rls.conn()` where auto-deref suffices
  (explicit_auto_deref) in create_audit_entry and list_audit_logs calls

+++ backend/servers/api-server/src/routes/migration.rs
@@ -893,7 +893,7 @@ async fn get_import_job_status(
 
     let progress_percent = if let Some(total) = job.total_rows {
         if total > 0 {
-            (job.processed_rows * 100 / total) as i32
+            job.processed_rows * 100 / total
         } else {
             0
         }
@@ -992,7 +992,7 @@ async fn cancel_import_job(
 
     let progress_percent = if let Some(total) = updated_job.total_rows {
         if total > 0 {
-            (updated_job.processed_rows * 100 / total) as i32
+            updated_job.processed_rows * 100 / total
         } else {
             0
         }
